fix(build): Xcode 16 compatibility for XNU flags and ApplePIODMA enums
Remove the internal-only clang flag dropped in Xcode 16.4, and fix
APIODMABitRange32 to use 64-bit shifts so enum initializers remain
valid integral constant expressions.

+++ src/Kernel/Extensions/IOPCIFamily/include/IOKit/apiodma/ApplePIODMADefinitions.h
@@ -10,7 +10,7 @@
 #include <IOKit/IOTypes.h>
 
 #define APIODMABit(bit)                  ((uint32_t)(1) << bit)
-#define APIODMABitRange32(start, end)    (~(((uint32_t)(1) << start) - 1) & (((uint32_t)(1) << end) | (((uint32_t)(1) << end) - 1)))
+#define APIODMABitRange32(start, end)    ((uint32_t)((~((((uint64_t)1) << (start)) - 1ULL)) & ((((uint64_t)1) << (end)) | ((((uint64_t)1) << (end)) - 1ULL))))
 #define APIODMABitRange64(start, end)    (~(((uint64_t)(1) << start) - 1) & (((uint64_t)(1) << end) | (((uint64_t)(1) << end) - 1)))
 #define APIODMABitRangePhase(start, end) (start)
